+++ networklibrary/src/main/java/com/eis/communication/network/CommandExecutor.java
@@ -1,9 +1,12 @@
 package com.eis.communication.network;
 
+import com.eis.communication.network.Commands.Command;
+
 /**
  * Class that manage all the sent commands.
- * It just let them executing.
+ * It just lets them executing.
  *
+ * @link https://refactoring.guru/design-patterns/command
  * @param <C> command type
  *
  * @author Edoardo Raimondi, idea by Marco Cognolato, Enrico cestaro
