+++ networklibrary/src/main/java/com/eis/communication/network/Commands/AddPeer.java
@@ -1,4 +1,27 @@
 package com.eis.communication.network.Commands;
 
-public class AddPeer {
+import com.eis.communication.network.SMSNetSubscribers;
+import com.eis.communication.network.SMSNetworkManager;
+import com.eis.smslibrary.SMSPeer;
+
+/**
+ * Command to add a peer to Subscribers
+ */
+public class AddPeer implements Command {
+
+    private SMSPeer peer;
+    private SMSNetworkManager networkManager;
+
+    public AddPeer(SMSPeer peer, SMSNetworkManager networkManager){
+        this.peer = peer;
+        this.networkManager = networkManager;
+    }
+
+    /**
+     * Add the peer to subscribers
+     */
+    public void execute(){
+        
+        //TODO broadcast
+    }
 }
