+++ networklibrary/src/main/java/com/eis/communication/network/SMSNetworkManager.java
@@ -1,0 +1,11 @@
+package com.eis.communication.network;
+
+/**
+ * The manager class of the network.
+ */
+public class SMSNetworkManager /*implements NetworkManager*/ {
+
+    private SMSNetSubscribers netSubscribers = new SMSNetSubscribers();
+    private SMSNetDictionary netDictionary = new SMSNetDictionary();
+
+}
